Move the ArrayBackend class into an array-api module.
