+++ odl/util/vectorization.py
@@ -11,16 +11,12 @@
 from __future__ import print_function, division, absolute_import
 from builtins import object
 from functools import wraps
-from dataclasses import dataclass
-from types import ModuleType
-from typing import Callable
 import numpy as np
 
 
 __all__ = ('is_valid_input_array', 'is_valid_input_meshgrid',
            'out_shape_from_meshgrid', 'out_shape_from_array',
-           'OptionalArgDecorator', 'vectorize',
-           'ArrayBackend', 'lookup_array_backend')
+           'OptionalArgDecorator', 'vectorize')
 
 
 def is_valid_input_array(x, ndim=None):
@@ -295,34 +291,6 @@ class _NumpyVectorizeWrapper(object):
         else:
             out[:] = self.vfunc(*x, **kwargs)
 
-
-
-_registered_array_backends = {}
-
-@dataclass
-class ArrayBackend:
-    impl: str
-    array_namespace: ModuleType
-    available_dtypes: dict[str, object]
-    array_type: type
-    array_constructor: Callable
-    identifier_of_dtype: Callable[object, str]
-    def __post_init__(self):
-        if self.impl in _registered_array_backends:
-            raise KeyError(f"An array-backend with the identifier {self.impl} is already registered."
-                          + " Every backend needs to have a unique identifier.")
-        _registered_array_backends[self.impl] = self
-    def get_dtype_identifier(self, **kwargs):
-        if 'array' in kwargs:
-            assert 'dtype' not in kwargs, 'array and dtype are multually exclusive parameters'
-            return self.identifier_of_dtype(kwargs['array'].dtype)
-        if 'dtype' in kwargs:
-            assert 'array' not in kwargs, 'array and dtype are multually exclusive parameters'
-            return self.identifier_of_dtype(kwargs['dtype'])
-        raise ValueError("Either 'array' or 'dtype' argument must be provided.")
-
-def lookup_array_backend(impl: str) -> ArrayBackend:
-    return _registered_array_backends[impl]
 
 
 if __name__ == '__main__':
